requirements.txt
flask
fastai==1.0.61
numpy
pillow
gunicorn